refactor: centralize pool match generation
